Folded ClientMessageFactory class into AuthClient class
