+++ jenjin-core-client/src/main/java/com/jenjinstudios/client/message/ClientMessageFactory.java
@@ -12,25 +12,4 @@ import com.jenjinstudios.core.io.MessageRegistry;
 public class ClientMessageFactory
 {
 
-    /**
-     * Generate a LogoutRequest message.
-     *
-     * @return The LogoutRequestMessage.
-     */
-    public static Message generateLogoutRequest() {
-        return MessageRegistry.getInstance().createMessage
-              ("LogoutRequest");
-    }
-
-    /**
-     * Generate a LoginRequest message.  This message will be encrypted if possible.
-     *
-     * @return The LoginRequest message.
-     */
-    public static Message generateLoginRequest(ClientUser user) {// Create the login request.
-        Message loginRequest = MessageRegistry.getInstance().createMessage("LoginRequest");
-        loginRequest.setArgument("username", user.getUsername());
-        loginRequest.setArgument("password", user.getPassword());
-        return loginRequest;
-    }
 }
