+++ jenjin-core-client/src/test/java/com/jenjinstudios/client/message/ClientMessageFactoryTest.java
@@ -1,5 +1,6 @@
 package com.jenjinstudios.client.message;
 
+import com.jenjinstudios.client.net.AuthClient;
 import com.jenjinstudios.client.net.ClientUser;
 import com.jenjinstudios.core.io.Message;
 import org.testng.Assert;
@@ -10,19 +11,19 @@ import org.testng.annotations.Test;
  */
 public class ClientMessageFactoryTest
 {
-	@Test
-	public void testGenerateLogoutRequest() {
-		ClientMessageFactory messageFactory = new ClientMessageFactory();
-		Message message = messageFactory.generateLogoutRequest();
-		Assert.assertEquals(message.name, "LogoutRequest");
-	}
+    @Test
+    public void testGenerateLogoutRequest() {
+        ClientMessageFactory messageFactory = new ClientMessageFactory();
+        Message message = AuthClient.generateLogoutRequest();
+        Assert.assertEquals(message.name, "LogoutRequest");
+    }
 
-	@Test
-	public void testGenerateLoginRequest() {
-		ClientMessageFactory messageFactory = new ClientMessageFactory();
-		Message message = messageFactory.generateLoginRequest(new ClientUser("Foo", "Bar"));
-		Assert.assertEquals(message.name, "LoginRequest");
-		Assert.assertEquals(message.getArgument("username"), "Foo");
-		Assert.assertEquals(message.getArgument("password"), "Bar");
-	}
+    @Test
+    public void testGenerateLoginRequest() {
+        ClientMessageFactory messageFactory = new ClientMessageFactory();
+        Message message = AuthClient.generateLoginRequest(new ClientUser("Foo", "Bar"));
+        Assert.assertEquals(message.name, "LoginRequest");
+        Assert.assertEquals(message.getArgument("username"), "Foo");
+        Assert.assertEquals(message.getArgument("password"), "Bar");
+    }
 }
